fixed crash-leading bug in checkers.c related to reallocs of size 0 and invalid frees.Fixed checkers_ai.c running forever on the minimaxr function and added shuffling function to try and make ai moves a bit more 'random'

+++ src/game/checkers.h
@@ -50,9 +50,9 @@ enum GameState {
 
 struct Checkers {
     struct {
-        int forceCapture;
-        int run;
-        int aiEnabled;
+        uint8_t forceCapture;
+        uint8_t run;
+        uint8_t aiEnabled;
     } flags;
     int turnsTotal;
     enum GameState state;
